genrb uses ustdio instead of local copies of ufile & ustdio
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/232



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@707 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/extra/ustdio/uscanf.c
@@ -526,11 +526,12 @@ u_scanf_scidbl_handler(UFILE             *stream,
 {
   int32_t        len;
   double        *num         = (double*) (args[0].ptrValue);
-  UNumberFormat        *scientificFormat, *genericFormat;
-  int32_t        scientificResult, genericResult;
-  int32_t        scientificParsePos = 0, genericParsePos = 0;
-  UErrorCode         scientificStatus = U_ZERO_ERROR;
-  UErrorCode         genericStatus = U_ZERO_ERROR;
+  UNumberFormat *scientificFormat, *genericFormat;
+  /*int32_t       scientificResult, genericResult;*/
+  double        scientificResult, genericResult;
+  int32_t       scientificParsePos = 0, genericParsePos = 0;
+  UErrorCode    scientificStatus = U_ZERO_ERROR;
+  UErrorCode    genericStatus = U_ZERO_ERROR;
   bool_t        useScientific;
